# api/swagger.yml
openapi: 3.1.0
info:
  title: Receipt Wrangler API.
  termsOfService:
    there are no TOS at this moment, use at your own risk we take no
    responsibility
  license:
    name: GPLV3
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  version: 5.0.0
servers:
  - url: /api
security:
  - bearerAuth: [ ]
  - apiKeyAuth: [ ]
paths:
...
  /dashboard/{dashboardId}:
    parameters:
      - in: path
        name: dashboardId
        schema:
          type: integer
        required: true
        description: Id of dashboard to operate on
    put:
      tags:
        - Dashboard
      summary: Update dashboard
      description: This will update a dashboard
      operationId: updateDashboard
      requestBody:
        description: Dashboard to update
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertDashboardCommand"
      responses:
        200:
          description: The updated dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
        400:
          $ref: "#/components/responses/BadRequest"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
        - apiKeyAuth: [ ]
    delete:
      tags:
        - Dashboard
      summary: Delete dashboard
      description: This will delete a dashboard by id
      operationId: deleteDashboard
      responses:
        200:
          description: The deleted dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
        - apiKeyAuth: [ ]
  /dashboard/:
    post:
      tags:
        - Dashboard
      summary: Create dashboard
      description: This will create a dashboard [SYSTEM USER]
      requestBody:
        description: Dashboard
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertDashboardCommand"
      operationId: createDashboard
      responses:
        200:
          description: The created dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
        - apiKeyAuth: [ ]
...
  /export:
    post:
      tags:
        - Export
      summary: Exports receipts
      description:
        This will export individual receipts [SYSTEM USER]
      parameters:
        - in: query
          name: format
          schema:
            $ref: "#/components/schemas/ExportFormat"
          required: true
        - in: query
          name: receiptIds
          schema:
            type: array
            items:
              type: integer
      operationId: exportReceiptsById
      responses:
        200:
          description: The csv stream
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
        - apiKeyAuth: [ ]
...
  /apiKey/:
    post:
      tags:
        - ApiKey
      summary: Create API key
      description: Create a new API key for the authenticated user
      operationId: createApiKey
      requestBody:
        description: API key details
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertApiKeyCommand"
      responses:
        200:
          description: The created API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyResult"
        400:
          $ref: "#/components/responses/BadRequest"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
        - apiKeyAuth: [ ]
  /apiKey/paged:
    post:
      tags:
        - ApiKey
      summary: Get paged API keys
      description: This will return paged API keys for the authenticated user or all API keys for admins
      operationId: getPagedApiKeys
      requestBody:
        description: Paging and sorting data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PagedApiKeyRequestCommand"
      responses:
        200:
          description: Paged API keys
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedData"
        400:
          $ref: "#/components/responses/BadRequest"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
        - apiKeyAuth: [ ]
  /apiKey/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: API key ID to update
    put:
      tags:
        - ApiKey
      summary: Update API key
      description: This will update an API key. Users can only update their own API keys.
      operationId: updateApiKey
      requestBody:
        description: API key details to update
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertApiKeyCommand"
      responses:
        200:
          $ref: "#/components/responses/Ok"
        400:
          $ref: "#/components/responses/BadRequest"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
        - apiKeyAuth: [ ]
    delete:
      tags:
        - ApiKey
      summary: Delete API key
      description: Delete an API key. Admins can delete any API key, non-admins can only delete their own API keys.
      operationId: deleteApiKey
      responses:
        200:
          $ref: "#/components/responses/Ok"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
        - apiKeyAuth: [ ]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: API Key in format v1.xxx.xxx.xxx
  responses:
    Ok:
      description: Request was successfully processed
    BadRequest:
      description: The request was malformed
...
    Forbidden:
      description: The request was not allowed
...
    Internal:
      description: There was an error processing the request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalErrorResponse'
  schemas:
...
    ApiKeyScope:
      description: Scope/permissions for API keys
      type: string
      enum:
        - "r"
        - "w"
        - "rw"
...
    GroupStatus:
      type: string
      enum:
        - "ACTIVE"
        - "ARCHIVED"
...
    SortDirection:
      type: string
      enum:
        - asc
        - desc
        - ""
    WidgetType:
      type: string
      enum:
        - "GROUP_SUMMARY"
        - "FILTERED_RECEIPTS"
        - "GROUP_ACTIVITY"
        - "PIE_CHART"
...
    SystemTaskStatus:
      type: string
      enum:
        - "SUCCEEDED"
        - "FAILED"
...
    ExportFormat:
      type: string
      enum:
        - "CSV"
...
    Activity:
      type: object
      required:
        - id
        - type
        - status
        - startedAt
        - endedAt
        - childSystemTasks
      properties:
        id:
          type: integer
        type:
          $ref: "#/components/schemas/SystemTaskType"
        status:
          $ref: "#/components/schemas/SystemTaskStatus"
        startedAt:
          type: string
        endedAt:
          type: string
        ranByUserId:
          type: integer
          format: uint64
        receiptId:
          type: integer
          format: uint64
        groupId:
          type: integer
          format: uint64
        canBeRestarted:
          type: boolean
...
    Category:
      type: object
      properties:
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        id:
          type: integer
          format: uint64
        name:
          type: string
          description: Name of the category
        description:
          type: string
          description: Description of the category
        updatedAt:
          type: string
          x-go-name: UpdatedAt
      description: Category to relate receipts to
...
    Group:
      required:
        - name
        - status
        - id
        - groupMembers
        - isAllGroup
        - groupReceiptSettings
      type: object
      properties:
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        groupSettings:
          $ref: "#/components/schemas/GroupSettings"
        groupReceiptSettings:
          $ref: "#/components/schemas/GroupReceiptSettings"
        groupMembers:
          type: array
          description: Members of the group
          items:
            $ref: "#/components/schemas/GroupMember"
        id:
          type: integer
          format: uint64
        isDefault:
          type: boolean
          description: Is default group (not used yet)
        name:
          type: string
          description: Name of the group
        isAllGroup:
          type: boolean
          description: Is all group for user
        status:
          type: string
          $ref: "#/components/schemas/GroupStatus"
        updatedAt:
          type: string
      description: Group in the system
...
    Receipt:
      required:
        - amount
        - date
        - groupId
        - name
        - paidByUserId
        - status
        - customFields
        - categories
        - tags
        - receiptItems
        - comments
        - id
      type: object
      properties:
        amount:
          type: string
          description: Receipt total amount
        categories:
          type: array
          description: Categories associated to receipt
          items:
            $ref: "#/components/schemas/Category"
        comments:
          type: array
          description: Comments associated to receipt
          items:
            $ref: "#/components/schemas/Comment"
        customFields:
          type: array
          description: Custom fields associated to receipt
          items:
            $ref: "#/components/schemas/CustomFieldValue"
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        date:
          type: string
          description: Receipt date
        groupId:
          type: integer
          description: Group foreign key
          format: uint64
        id:
          type: integer
          format: uint64
        imageFiles:
          type: array
          description: Files associated to receipt
          items:
            $ref: "#/components/schemas/FileData"
        name:
          type: string
          description: Receipt name
        paidByUserId:
          type: integer
          description: User paid foreign key
          format: uint64
        receiptItems:
          type: array
          description: Items associated to receipt
          items:
            $ref: "#/components/schemas/Item"
        resolvedDate:
          type: string
          description: Date resolved
        status:
          $ref: "#/components/schemas/ReceiptStatus"
        tags:
          type: array
          description: Tags associated to receipt
          items:
            $ref: "#/components/schemas/Tag"
        updatedAt:
          type: string
        createdByString:
          type: string
          description: Created by string, which is anything that is not a user
      description: Receipt
    Tag:
      required:
        - name
      type: object
      properties:
        createdAt:
          type: string
          x-go-name: CreatedAt
        createdBy:
          type: integer
          format: uint64
        id:
          type: integer
          format: uint64
        name:
          type: string
          description: Tag name
        description:
          type: string
          description: Tag description
        updatedAt:
          type: string
      description: Tag to relate receipts to
    Dashboard:
      required:
        - id
        - name
        - userId
      type: object
      properties:
        createdAt:
          type: string
          x-go-name: CreatedAt
        createdBy:
          type: integer
          format: uint64
        id:
          type: integer
          format: uint64
        name:
          type: string
          description: Dashboard name
        groupId:
          type: integer
          format: uint64
          description: Group foreign key
        userId:
          type: integer
          format: uint64
          description: User foreign key
        updatedAt:
          type: string
        widgets:
          type: array
          description: Widgets associated to dashboard
          items:
            $ref: "#/components/schemas/Widget"
      description: Dashboard for a user
    Widget:
      required:
        - id
        - dashboardId
      type: object
      properties:
        createdAt:
          type: string
          x-go-name: CreatedAt
        createdBy:
          type: integer
          format: uint64
        id:
          type: integer
          format: uint64
        name:
          type: string
          description: Widget name
        dashboardId:
          type: integer
          format: uint64
          description: Dashboard foreign key
        updatedAt:
          type: string
        widgetType:
          type: string
          $ref: "#/components/schemas/WidgetType"
          description: Type of widget
        configuration:
          type: object
          additionalProperties: { }
          description: Configuration of widget
      description: Widget related to a user's dashboard
...
    TagView:
      required:
        - id
        - name
        - numberOfReceipts
      type: object
      properties:
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        id:
          type: integer
          format: uint64
        name:
          type: string
          description: Name of the tag
        description:
          type: string
          description: Description of the tag
        updatedAt:
          type: string
        numberOfReceipts:
          type: integer
          description: Number of receipts associated with this tag
      description: Tag to relate receipts to
...
    PagedRequestCommand:
      required:
        - page
        - pageSize
      type: object
      properties:
        page:
          type: integer
          description: Page number
        pageSize:
          type: integer
          description: Number of records per page
        orderBy:
          type: string
          description: field to order on
        sortDirection:
          type: string
          $ref: "#/components/schemas/SortDirection"
...
    PagedData:
      required:
        - data
        - totalCount
      type: object
      properties:
        data:
          type: array
          items:
            anyOf:
              - $ref: "#/components/schemas/Receipt"
              - $ref: "#/components/schemas/Category"
              - $ref: "#/components/schemas/Tag"
              - $ref: "#/components/schemas/Prompt"
              - $ref: "#/components/schemas/Group"
              - $ref: "#/components/schemas/TagView"
              - $ref: "#/components/schemas/SystemTask"
              - $ref: "#/components/schemas/ReceiptProcessingSettings"
              - $ref: "#/components/schemas/SystemEmail"
              - $ref: "#/components/schemas/Activity"
              - $ref: "#/components/schemas/CustomField"
        totalCount:
          type: integer
...
    UpsertDashboardCommand:
      type: object
      required:
        - name
        - groupId
      properties:
        name:
          type: string
          description: Dashboard name
        groupId:
          type: string
          description: Group foreign key
        widgets:
          type: array
          description: Widgets associated to dashboard
          items:
            $ref: "#/components/schemas/UpsertWidgetCommand"
    UpsertWidgetCommand:
      type: object
      required:
        - widgetType
      properties:
        name:
          type: string
          description: Widget name
        widgetType:
          type: string
          $ref: "#/components/schemas/WidgetType"
          description: Type of widget
        configuration:
          type: object
          additionalProperties: { }
          description: Configuration of widget
...
    SystemEmail:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          properties:
            host:
              type: string
              description: IMAP host
            port:
              type: string
              description: IMAP port
            username:
              type: string
              description: IMAP username
            password:
              type: string
              description: IMAP password
            useStartTLS:
              type: boolean
              description: Whether to use STARTTLS
...
    SystemTask:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          properties:
            type:
              $ref: "#/components/schemas/SystemTaskType"
            status:
              $ref: "#/components/schemas/SystemTaskStatus"
            startedAt:
              type: string
            endedAt:
              type: string
            associatedEntityId:
              type: integer
            associatedEntityType:
              $ref: "#/components/schemas/AssociatedEntityType"
            ranByUserId:
              type: integer
              format: uint64
            receiptId:
              type: integer
              format: uint64
            groupId:
              type: integer
              format: uint64
            resultDescription:
              type: string
            apiKeyId:
              type: string
              nullable: true
            childSystemTasks:
              type: array
              items:
                $ref: "#/components/schemas/SystemTask"
...
    ReceiptProcessingSettings:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          properties:
            name:
              type: string
              description: Name of the settings
            description:
              type: string
              description: Description of the settings
            aiType:
              $ref: "#/components/schemas/AiType"
            url:
              type: string
              description: URL for custom endpoints
            key:
              type: string
              description: Key for endpoints that require authentication
            model:
              type: string
              description: LLM model
            isVisionModel:
              type: boolean
              description: Is vision model
            enforceJsonResponseFormat:
              type: boolean
              description: Enforce JSON response format on the LLM provider. Disable if the provider does not support this flag.
            ocrEngine:
              $ref: "#/components/schemas/OcrEngine"
            prompt:
              $ref: "#/components/schemas/Prompt"
            promptId:
              type: integer
              description: Prompt foreign key
...
    Prompt:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          required:
            - name
            - prompt
          properties:
            name:
              type: string
              description: Prompt name
            description:
              type: string
              description: Prompt description
            prompt:
              type: string
              description: Prompt text
...
    UpsertApiKeyCommand:
      type: object
      required:
        - name
        - scope
      properties:
        name:
          type: string
          description: API key name
        description:
          type: string
          description: API key description
        scope:
          $ref: "#/components/schemas/ApiKeyScope"
    ApiKeyResult:
      type: object
      required:
        - key
      properties:
        key:
          type: string
          description: The generated API key
    PagedApiKeyRequestCommand:
      allOf:
        - $ref: "#/components/schemas/PagedRequestCommand"
        - type: object
          properties:
            filter:
              $ref: "#/components/schemas/ApiKeyFilter"
    ApiKeyFilter:
      type: object
      properties:
        associatedApiKeys:
          $ref: "#/components/schemas/AssociatedApiKeys"
    AssociatedApiKeys:
      type: string
      enum:
        - "MINE"
        - "ALL"
...
    InternalErrorResponse:
      type: object
      required:
        - errorMsg
      properties:
        errorMsg:
          type: string
          description: Error message
...